+++ led_strip/src/led_strip_rmt_encoder.h
@@ -7,6 +7,7 @@
 
 #include <stdint.h>
 #include "driver/rmt_encoder.h"
+#include "led_strip_types.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -17,6 +18,7 @@ extern "C" {
  */
 typedef struct {
     uint32_t resolution; /*!< Encoder resolution, in Hz */
+    led_model_t led_model;
 } led_strip_encoder_config_t;
 
 /**