+++ led_strip/CHANGELOG.md
@@ -12,3 +12,12 @@
 - Support various RMT clock sources
 - Acquire and release the power management lock before and after each refresh
 - New driver flag: `invert_out` which can invert the led control signal by hardware
+
+## 2.2.0
+
+- Support for 4 components RGBW leds (SK6812):
+  - in led_strip_config_t new fields
+      led_pixel_format, controlling byte format (LED_PIXEL_FORMAT_GRB, LED_PIXEL_FORMAT_GRBW)
+      led_model, used to configure bit timing (LED_MODEL_WS2812, LED_MODEL_SK6812)
+  - new API led_strip_set_pixel_rgbw
+  - new interface type set_pixel_rgbw